+++ pymk/tests/dependency.py
@@ -3,9 +3,9 @@ import pymk.error as Perror
 from pymk import extra
 from pymk.tests.base import PymkTestCase
 
-class FileDoesNotExistsConditionTest(PymkTestCase):
+class FileDoesNotExistsDependencyTest(PymkTestCase):
     def test_make(self):
-        self._template('one_task_condition_1', 'mkfile.py')
+        self._template('one_task_dependency_1', 'mkfile.py')
         self._import_mkfile()
 
         self._args.task.append('task_2')
@@ -14,18 +14,18 @@ class FileDoesNotExistsConditionTest(PymkTestCase):
         self._pymk_runtask(['task_2', 'task_2', 'task_2'])
 
     def test_uptodate(self):
-        self._template('one_task_condition_1', 'mkfile.py')
-        self._template('one_task_condition_1', 'test.txt')
+        self._template('one_task_dependency_1', 'mkfile.py')
+        self._template('one_task_dependency_1', 'test.txt')
         self._import_mkfile()
 
         self._add_task('task_2')
         self._pymk_runtask([])
         self._pymk_runtask([])
 
-class FileChangedConditionTest(PymkTestCase):
+class FileChangedDependencyTest(PymkTestCase):
 
-    def test_FileChanged_condition_make_fail(self):
-        self._template('one_task_condition_2', 'mkfile.py')
+    def test_FileChanged_dependency_make_fail(self):
+        self._template('one_task_dependency_2', 'mkfile.py')
         self._import_mkfile()
 
         self._add_task('task_3')
@@ -33,8 +33,8 @@ class FileChangedConditionTest(PymkTestCase):
         self.assertRaises(Perror.TaskMustHaveOutputFile, self._pymk)
         self._check_output_file([])
 
-    def test_FileChanged_condition_make_fail_2(self):
-        self._template('one_task_condition_2', 'mkfile.py')
+    def test_FileChanged_dependency_make_fail_2(self):
+        self._template('one_task_dependency_2', 'mkfile.py')
         self.touch('test.txt', None)
         self._import_mkfile()
 
@@ -43,8 +43,8 @@ class FileChangedConditionTest(PymkTestCase):
         self.assertRaises(Perror.TaskMustHaveOutputFile, self._pymk)
         self._check_output_file([])
 
-    def test_FileChanged_condition_make_faile_3(self):
-        self._template('one_task_condition_3', 'mkfile.py')
+    def test_FileChanged_dependency_make_faile_3(self):
+        self._template('one_task_dependency_3', 'mkfile.py')
         self._import_mkfile()
 
         self._add_task('task_4')
@@ -52,8 +52,8 @@ class FileChangedConditionTest(PymkTestCase):
         self.assertRaises(Perror.CouldNotCreateFile, self._pymk)
         self._check_output_file([])
 
-    def test_FileChanged_condition_make_once(self):
-        self._template('one_task_condition_3', 'mkfile.py')
+    def test_FileChanged_dependency_make_once(self):
+        self._template('one_task_dependency_3', 'mkfile.py')
         self._import_mkfile()
 
         self._add_task('task_4')
@@ -62,8 +62,8 @@ class FileChangedConditionTest(PymkTestCase):
         self._pymk_runtask(['task_4'])
         self._pymk_runtask(['task_4'])
 
-    def test_FileChanged_condition_make_twice(self):
-        self._template('one_task_condition_3', 'mkfile.py')
+    def test_FileChanged_dependency_make_twice(self):
+        self._template('one_task_dependency_3', 'mkfile.py')
         self._import_mkfile()
 
         self._add_task('task_4')
@@ -77,9 +77,9 @@ class FileChangedConditionTest(PymkTestCase):
         sleep(0.001)
         self._pymk_runtask(['task_4', 'task_4'])
 
-class AlwaysRebuildConditionTest(PymkTestCase):
+class AlwaysRebuildDependencyTest(PymkTestCase):
     def test_success(self):
-        self._template('three_task_condition_always1', 'mkfile.py')
+        self._template('three_task_dependency_always1', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_16a')
 
@@ -88,7 +88,7 @@ class AlwaysRebuildConditionTest(PymkTestCase):
         self._pymk_runtask(['task_16c', 'task_16b', 'task_16a', 'task_16a', 'task_16a'])
 
     def test_dependency_rebuild(self):
-        self._template('three_task_condition_always1', 'mkfile.py')
+        self._template('three_task_dependency_always1', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_16a')