+++ pymk/tests/task.py
@@ -30,7 +30,7 @@ class TaskTest(PymkTestCase):
         self.assertEqual(self._pymk(), 'run default')
         self._check_output_file(['task_0'])
 
-    def test_no_condition(self):
+    def test_no_dependency(self):
         self._template('one_task', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_1')
@@ -39,7 +39,7 @@ class TaskTest(PymkTestCase):
         self._pymk_runtask(['task_1', 'task_1'])
         self._pymk_runtask(['task_1', 'task_1', 'task_1'])
 
-    def test_no_condition_with_outputfile(self):
+    def test_no_dependency_with_outputfile(self):
         self._template('one_task_with_outputfile', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_5')
@@ -49,9 +49,9 @@ class TaskTest(PymkTestCase):
         self.touch('a.out')
         self._pymk_runtask(['task_5'])
 
-class TaskConditionFileExistsTest(PymkTestCase):
+class TaskDependencyFileExistsTest(PymkTestCase):
     def test_make_no_outputfile(self):
-        self._template('three_task_condition_exists2', 'mkfile.py')
+        self._template('three_task_dependency_exists2', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_9a')
 
@@ -61,7 +61,7 @@ class TaskConditionFileExistsTest(PymkTestCase):
         self._pymk_runtask(['task_9c', 'task_9b', 'task_9a'])
 
     def test_make_no_outputfile_fail1(self):
-        self._template('three_task_condition_exists3', 'mkfile.py')
+        self._template('three_task_dependency_exists3', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_10a')
 
@@ -76,7 +76,7 @@ class TaskConditionFileExistsTest(PymkTestCase):
         self._check_output_file([])
 
     def test_make_no_outputfile_fail2(self):
-        self._template('three_task_condition_exists4', 'mkfile.py')
+        self._template('three_task_dependency_exists4', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_11a')
 
@@ -91,7 +91,7 @@ class TaskConditionFileExistsTest(PymkTestCase):
         self._check_output_file([])
 
     def test_make(self):
-        self._template('two_task_condition_exists1', 'mkfile.py')
+        self._template('two_task_dependency_exists1', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_6a')
 
@@ -101,7 +101,7 @@ class TaskConditionFileExistsTest(PymkTestCase):
         self._pymk_runtask(['task_6b', 'task_6a'])
 
     def test_make_2(self):
-        self._template('three_task_condition_exists1', 'mkfile.py')
+        self._template('three_task_dependency_exists1', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_7a')
 
@@ -122,7 +122,7 @@ class TaskConditionFileExistsTest(PymkTestCase):
         self._pymk_runtask(['task_7c', 'task_7b', 'task_7a', 'task_7b'])
 
     def test_uptodate(self):
-        self._template('two_task_condition_exists1', 'mkfile.py')
+        self._template('two_task_dependency_exists1', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_6a')
         self.touch('b.out', None)
@@ -131,9 +131,9 @@ class TaskConditionFileExistsTest(PymkTestCase):
         self._pymk_runtask(['task_6a'])
         self._pymk_runtask(['task_6a'])
 
-class TaskConditionFileChangedTest(PymkTestCase):
+class TaskDependencyFileChangedTest(PymkTestCase):
     def test_make_once(self):
-        self._template('two_task_condition_changed1', 'mkfile.py')
+        self._template('two_task_dependency_changed1', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_8a')
 
@@ -142,7 +142,7 @@ class TaskConditionFileChangedTest(PymkTestCase):
         self._pymk_runtask(['task_8b', 'task_8a'])
 
     def test_make_twice(self):
-        self._template('two_task_condition_changed1', 'mkfile.py')
+        self._template('two_task_dependency_changed1', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_8a')
 
@@ -152,7 +152,7 @@ class TaskConditionFileChangedTest(PymkTestCase):
         self._pymk_runtask(['task_8b', 'task_8a', 'task_8a'])
 
     def test_make_when_depedency_lost(self):
-        self._template('two_task_condition_changed1', 'mkfile.py')
+        self._template('two_task_dependency_changed1', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_8a')
 
@@ -162,7 +162,7 @@ class TaskConditionFileChangedTest(PymkTestCase):
         self._pymk_runtask(['task_8b', 'task_8a', 'task_8b', 'task_8a'])
 
     def test_make_three_tasks(self):
-        self._template('three_task_condition_changed1', 'mkfile.py')
+        self._template('three_task_dependency_changed1', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_12a')
 
@@ -182,7 +182,7 @@ class TaskConditionFileChangedTest(PymkTestCase):
         self._pymk_runtask(['task_12c', 'task_12b', 'task_12a', 'task_12a', 'task_12b', 'task_12a', 'task_12b', 'task_12a'])
 
     def test_make_three_tasks_fail1(self):
-        self._template('three_task_condition_changed2', 'mkfile.py')
+        self._template('three_task_dependency_changed2', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_13a')
 
@@ -202,7 +202,7 @@ class TaskConditionFileChangedTest(PymkTestCase):
         self._check_output_file(['task_13c', 'task_13b', 'task_13b', 'task_13b'])
 
     def test_make_three_tasks_fail2(self):
-        self._template('three_task_condition_changed3', 'mkfile.py')
+        self._template('three_task_dependency_changed3', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_14a')
 
@@ -222,7 +222,7 @@ class TaskConditionFileChangedTest(PymkTestCase):
         self._check_output_file(['task_14c'])
 
     def test_make_three_tasks_fail3(self):
-        self._template('three_task_condition_changed4', 'mkfile.py')
+        self._template('three_task_dependency_changed4', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_15a')
 
@@ -242,7 +242,7 @@ class TaskConditionFileChangedTest(PymkTestCase):
         self._check_output_file(['task_15c', 'task_15c', 'task_15c', 'task_15c'])
 
     def test_uptodate(self):
-        self._template('two_task_condition_changed1', 'mkfile.py')
+        self._template('two_task_dependency_changed1', 'mkfile.py')
         self._import_mkfile()
         self._add_task('task_8a')
         self.touch('b.out', None)